+++ test/Lib/hu.js
@@ -4366,19 +4366,36 @@
   const util = create( null );
 
   assign( util, {
+    /** 绑定事件 */
     addEvent: addEventListener,
+    /** 移除事件 */
     removeEvent: removeEventListener,
+    /** 对象遍历方法 */
     each,
+    /** 将值转为字符串形式 */
+    toString: toString$1,
+    /** 判断传入对象是否是纯粹的对象 */
     isPlainObject,
+    /** 判断传入对象是否是一个空对象 */
     isEmptyObject,
+    /** 判断传入对象是否是原始对象 */
     isPrimitive,
+    /** 判断传入对象是否可迭代 */
+    isIterable,
+    /** 判断传入的两个值是否相等 */
     isEqual,
+    /** 判断传入的两个值是否不相等 */
+    isNotEqual,
+    /** 判断传入对象是否是 String 类型 */
     isString,
+    /** 判断传入对象是否是 Object 类型且不为 null */
     isObject,
+    /** 判断传入对象是否是 Function 类型 */
     isFunction,
+    /** 判断传入对象是否是 Symbol 类型 */
     isSymbol,
-    uid: uid$1,
-    cached
+    /** 返回一个字符串 UID */
+    uid: uid$1
   });
 
   function directive( name, directive ){